+++ apps/rosen-service2/src/types/chains.ts
@@ -1,15 +1,11 @@
 import { NETWORKS_KEYS } from '@rosen-ui/constants';
 
 import { configs } from '../configs';
-import { BITCOIN_RUNES_NETWORKS_KEY } from '../constants';
 
-// TODO: implement Bitcoin-Runes support later
-export type ChainChoices = Exclude<
-  (typeof NETWORKS_KEYS)[number],
-  typeof BITCOIN_RUNES_NETWORKS_KEY
->;
+export type ChainChoices = (typeof NETWORKS_KEYS)[number];
 
 export type ChainsKeys = keyof (typeof configs)['chains'];
+export type ChainsWithScanner = Exclude<ChainsKeys, 'bitcoin-runes'>;
 
 export interface ChainConfigs {
   addresses: {