+++ packages/rosen-statistics-entity/lib/actions/index.ts
@@ -2,3 +2,5 @@ export { MetricAction } from './metricAction';
 export { LockedAssetsMetricAction } from './lockedAssetsMetricAction';
 export { EventCountMetricAction } from './eventCountMetricAction';
 export { UserEventMetricAction } from './userEventMetricAction';
+export { WatcherCountMetricAction } from './watcherCountMetricAction';
+export { BridgeMetricsAction } from './bridgeMetricsAction';