+++ packages/rosen-statistics/lib/fetchers/index.ts
@@ -1,0 +1,1 @@
+export { NodeBoxFetcher } from './nodeBoxFetcher';