+++ packages/rosen-statistics/lib/index.ts
@@ -3,4 +3,8 @@ export {
   lockedAssetsMetric,
   eventCountMetric,
   userEventMetric,
+  watcherCountMetric,
+  bridgeFeeMetric,
 } from './jobs';
+export { WatcherCountConfig } from './types';
+export { NodeBoxFetcher } from './fetchers';