+++ packages/rosen-statistics/lib/jobs/index.ts
@@ -2,3 +2,5 @@ export { generalMetrics } from './generalMetrics';
 export { lockedAssetsMetric } from './lockedAssetsMetric';
 export { eventCountMetric } from './eventCountMetric';
 export { userEventMetric } from './userEventMetric';
+export { watcherCountMetric } from './watcherCountMetric';
+export { bridgeFeeMetric } from './bridgeFeeMetric';