+++ packages/rosen-statistics/lib/types/index.ts
@@ -1,0 +1,5 @@
+export {
+  AggregatedBridgeFeeData,
+  BridgeFeeCalculationResult,
+} from './bridgeFee';
+export { WatcherCountConfig, WatcherCountResult } from './watcherCount';