+++ packages/rosen-statistics/lib/utils/index.ts
@@ -1,3 +1,5 @@
+export { calculateBridgeFees } from './bridgeFeeCalculator';
+export { calculateWatcherCounts } from './calculateWatcherCounts ';
 export {
   getDecimalString,
   getNonDecimalString,
@@ -5,3 +7,4 @@ export {
   scientificToString,
   multiplyByPowerOfTen,
 } from './decimals';
+export { getRegisterValue } from './getRegisterValue';