+++ apps/rosen-service/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @rosen-bridge/rosen-service
 
+## 4.3.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@1.0.2
+  - @rosen-bridge/bitcoin-observation-extractor@7.0.3
+  - @rosen-bridge/bitcoin-runes-observation-extractor@2.0.3
+  - @rosen-bridge/bitcoin-scanner@1.0.2
+  - @rosen-bridge/cardano-observation-extractor@2.0.3
+  - @rosen-bridge/cardano-scanner@2.0.2
+  - @rosen-bridge/ergo-observation-extractor@1.0.3
+  - @rosen-bridge/ergo-scanner@1.0.2
+  - @rosen-bridge/evm-observation-extractor@6.0.3
+  - @rosen-bridge/evm-scanner@1.0.2
+  - @rosen-bridge/watcher-data-extractor@13.0.4
+  - @rosen-ui/data-source@0.2.4
+  - @rosen-ui/asset-calculator@2.3.2
+
 ## 4.3.3
 
 ### Patch Changes