+++ networks/base/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosen-network/base
 
+## 0.5.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/address-codec@1.2.0
+  - @rosen-ui/types@0.5.0
+  - @rosen-ui/utils@1.0.3
+
 ## 0.5.2
 
 ### Patch Changes