+++ networks/ergo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosen-network/ergo
 
+## 2.5.6
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/icons@3.6.0
+  - @rosen-ui/types@0.5.0
+  - @rosen-network/base@0.5.3
+
 ## 2.5.5
 
 ### Patch Changes