+++ networks/evm/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rosen-network/evm
 
+## 0.3.7
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/address-codec@1.2.0
+  - @rosen-ui/types@0.5.0
+
 ## 0.3.6
 
 ### Patch Changes